Refactor toolchain handling and use edalize for icestorm
This commit moves toolchain, utils and icestorm targets
(arachne and nextpnr) out of main script. This way code
is more readable.

Also icestorm targets are now using edalize.
