+++ utils.py
@@ -1,0 +1,16 @@
+import time
+
+class Timed():
+    def __init__(self, t, name):
+        self.t = t
+        self.name = name
+        self.start = None
+
+    def __enter__(self):
+        self.start = time.time()
+
+    def __exit__(self, type, value, traceback):
+        end = time.time()
+        self.t.add_runtime(self.name, end - self.start)
+
+
